Move router into utilities
