--- core/router/build.gradle
@@ -1,15 +1,0 @@
-/*
- * Copyright 2016 Year4000. All Rights Reserved.
- */
-
-plugins {
-    id 'java'
-}
-
-apply from: utilities.file('gradle/utilities.gradle')
-
-dependencies {
-    compile utilities.project('core')
-    compile 'io.netty:netty-all:5.0.0.Alpha2'
-}
-
