+++ core/src/main/java/net/year4000/utilities/net/router/Handle.java
@@ -2,14 +2,16 @@
  * Copyright 2015 Year4000. All Rights Reserved.
  */
 
-package net.year4000.utilities.router;
+package net.year4000.utilities.net.router;
 
 import io.netty.handler.codec.http.HttpRequest;
 import io.netty.handler.codec.http.HttpResponse;
 import net.year4000.utilities.value.TypeValue;
 
+import java.io.IOException;
+
 @FunctionalInterface
 public interface Handle<T> {
     /** Handle the content type from the client */
-    T handle(HttpRequest request, HttpResponse response, TypeValue... args) throws RouteHandleException;
+    T handle(HttpRequest request, HttpResponse response, TypeValue... args) throws RouteHandleException, IOException;
 }
