+++ core/src/main/java/net/year4000/utilities/net/router/RouteHandleException.java
@@ -1,4 +1,4 @@
-package net.year4000.utilities.router;
+package net.year4000.utilities.net.router;
 
 import net.year4000.utilities.Conditions;
