+++ core/src/main/java/net/year4000/utilities/net/router/RoutingManager.java
@@ -1,4 +1,4 @@
-package net.year4000.utilities.router;
+package net.year4000.utilities.net.router;
 
 import com.google.common.collect.ImmutableCollection;
 import com.google.common.collect.ImmutableList;
