+++ core/src/main/java/net/year4000/utilities/net/router/pipline/HttpInitializer.java
@@ -2,14 +2,14 @@
  * Copyright 2015 Year4000. All Rights Reserved.
  */
 
-package net.year4000.utilities.router.pipline;
+package net.year4000.utilities.net.router.pipline;
 
 import io.netty.channel.Channel;
 import io.netty.channel.ChannelInitializer;
 import io.netty.channel.ChannelPipeline;
 import io.netty.handler.codec.http.HttpServerCodec;
 import net.year4000.utilities.Conditions;
-import net.year4000.utilities.router.Router;
+import net.year4000.utilities.net.router.Router;
 
 public class HttpInitializer extends ChannelInitializer<Channel> {
     private final Router router;
@@ -27,8 +27,7 @@ public class HttpInitializer extends ChannelInitializer<Channel> {
 
     @Override
     protected void initChannel(Channel channel) throws Exception {
-        System.err.println("init");
-        channel.pipeline().firstContext().attr(Router.ATTRIBUTE_KEY).set(router);
+        channel.attr(Router.ATTRIBUTE_KEY).set(router);
         addHandlers(channel.pipeline());
     }
 }
