+++ settings.gradle
@@ -14,4 +14,3 @@ include 'core:sdk'
 include 'sponge'
 include 'bungee'
 include 'spigot'
-include 'core:router'
